CleanUp and remove winapi call from explorip to put them in ManagedShell.Interop

+++ Libraries/ManagedShell/ManagedShell.AppBar/AppBarManager.cs
@@ -109,7 +109,7 @@ public class AppBarManager : IDisposable
         // apparently the TaskBars like to pop up when AppBars change
         if (_explorerHelper.HideExplorerTaskbar)
         {
-            _explorerHelper.SetSecondaryTaskbarVisibility((int)SWP.SWP_HIDEWINDOW);
+            _explorerHelper.SetSecondaryTaskbarVisibility(SWP.SWP_HIDEWINDOW);
         }
     }
 
@@ -126,7 +126,7 @@ public class AppBarManager : IDisposable
         // apparently the TaskBars like to pop up when AppBars change
         if (_explorerHelper.HideExplorerTaskbar)
         {
-            _explorerHelper.SetSecondaryTaskbarVisibility((int)SWP.SWP_HIDEWINDOW);
+            _explorerHelper.SetSecondaryTaskbarVisibility(SWP.SWP_HIDEWINDOW);
         }
     }
